+++ src/lightning/fabric/plugins/precision/__init__.py
@@ -16,15 +16,15 @@ from lightning.fabric.plugins.precision.deepspeed import DeepSpeedPrecision
 from lightning.fabric.plugins.precision.double import DoublePrecision
 from lightning.fabric.plugins.precision.fsdp import FSDPPrecision
 from lightning.fabric.plugins.precision.precision import Precision
-from lightning.fabric.plugins.precision.tpu import TPUPrecision
-from lightning.fabric.plugins.precision.tpu_bf16 import TPUBf16Precision
+from lightning.fabric.plugins.precision.xla import XLAPrecision
+from lightning.fabric.plugins.precision.xlabf16 import XLABf16Precision
 
 __all__ = [
     "DeepSpeedPrecision",
     "DoublePrecision",
     "MixedPrecision",
     "Precision",
-    "TPUPrecision",
-    "TPUBf16Precision",
+    "XLAPrecision",
+    "XLABf16Precision",
     "FSDPPrecision",
 ]